drivers: disk: sdmmc: support idma with cache operations
On platforms with DCACHE, dma operation requires additional
cache maintenance operations to sync the data in cache with
RAM where dma happens. This patch adds the required cache
maintenance operation for read/write when dma is involved
and DCACHE is enabled. This is the case with STM32H7 and
other SoCs with DCACHE.

Signed-off-by: Murali Karicheri <murali.karicheri@sandc.com>